feat: module improvements
