+++ package.json
@@ -1,35 +1,35 @@
 {
   "name": "@nuxtjs/strapi",
   "version": "1.0.0",
-  "description": "Strapi module for Nuxt 3",
+  "description": "Strapi module for Nuxt",
   "repository": "nuxt-community/strapi-module",
   "license": "MIT",
-  "exports": {
-    ".": {
-      "require": "./dist/module.cjs",
-      "import": "./dist/module.mjs"
-    }
-  },
-  "main": "dist/module.mjs",
-  "module": "dist/module.mjs",
-  "types": "dist/module.d.ts",
+  "exports": "./dist/module.mjs",
+  "types": "./dist/module.d.ts",
   "files": [
     "dist"
   ],
   "scripts": {
     "build": "unbuild",
-    "dev": "nuxt dev example",
+    "dev": "nuxi dev example",
     "lint": "eslint --ext .js,.ts,.vue .",
     "prepublishOnly": "yarn build",
     "release": "yarn lint && standard-version && git push --follow-tags"
   },
+  "dependencies": {
+    "@nuxt/kit": "npm:@nuxt/kit-edge@latest",
+    "defu": "^5.0.0",
+    "pathe": "^0.2.0"
+  },
   "devDependencies": {
+    "@nuxt/schema": "npm:@nuxt/schema-edge@latest",
     "@nuxtjs/eslint-config-typescript": "^8.0.0",
     "eslint": "^8.5.0",
-    "nuxt3": "^3.0.0-27334958.ca9761d",
+    "nuxt3": "latest",
+    "ohmyfetch": "^0.4.11",
     "standard-version": "^9.3.2",
     "typescript": "^4.5.4",
-    "unbuild": "0.6.3"
+    "unbuild": "^0.6.7"
   },
   "publishConfig": {
     "access": "public"
