+++ src/runtime/composables/useStrapi4.ts
@@ -1,34 +1,23 @@
-import type { Strapi3RequestParams } from '../types/v3'
+import type { Strapi4RequestParams } from '../../types/v4'
 import { useStrapiVersion } from './useStrapiVersion'
 import { useStrapiClient } from './useStrapiClient'
 
-export const useStrapi3 = () => {
+export const useStrapi4 = () => {
   const client = useStrapiClient()
   const version = useStrapiVersion()
-  if (version !== 'v3') {
+  if (version !== 'v4') {
     // eslint-disable-next-line no-console
-    console.warn('useStrapi3 is only available for v3')
-  }
-
-  /**
-   * Count {content-type} entries
-   *
-   * @param  {string} contentType - Content type's name pluralized
-   * @param  {Strapi3RequestParams} params? - Query parameters
-   * @returns Promise<number>
-   */
-  const count = (contentType: string, params?: Strapi3RequestParams): Promise<number> => {
-    return client(`/${contentType}/count`, { method: 'GET', params })
+    console.warn('useStrapi4 is only available for v4')
   }
 
   /**
    * Get a list of {content-type} entries
    *
    * @param  {string} contentType - Content type's name pluralized
-   * @param  {Strapi3RequestParams} params? - Query parameters
+   * @param  {Strapi4RequestParams} params? - Query parameters
    * @returns Promise<T>
    */
-  const find = <T>(contentType: string, params?: Strapi3RequestParams): Promise<T> => {
+  const find = <T>(contentType: string, params?: Strapi4RequestParams): Promise<T> => {
     return client(`/${contentType}`, { method: 'GET', params })
   }
 
@@ -37,10 +26,10 @@ export const useStrapi3 = () => {
    *
    * @param  {string} contentType - Content type's name pluralized
    * @param  {string|number} id - ID of entry
-   * @param  {Strapi3RequestParams} params? - Query parameters
+   * @param  {Strapi4RequestParams} params? - Query parameters
    * @returns Promise<T>
    */
-  const findOne = <T>(contentType: string, id: string | number, params?: Strapi3RequestParams): Promise<T> => {
+  const findOne = <T>(contentType: string, id: string | number, params?: Strapi4RequestParams): Promise<T> => {
     return client(`/${contentType}/${id}`, { method: 'GET', params })
   }
 
@@ -52,7 +41,7 @@ export const useStrapi3 = () => {
    * @returns Promise<T>
    */
   const create = <T>(contentType: string, data: Partial<T>): Promise<T> => {
-    return client(`/${contentType}`, { method: 'POST', body: data })
+    return client(`/${contentType}`, { method: 'POST', body: { data } })
   }
 
   /**
@@ -63,15 +52,16 @@ export const useStrapi3 = () => {
    * @param  {Record<string, any>} data - Form data
    * @returns Promise<T>
    */
-  const update = <T>(contentType: string, id?: string | number | Partial<T>, data?: Partial<T>): Promise<T> => {
+  const update = <T>(contentType: string, id: string | number | Partial<T>, data?: Partial<T>): Promise<T> => {
     if (typeof id === 'object') {
       data = id
+      // @ts-ignore
       id = undefined
     }
 
     const path = [contentType, id].filter(Boolean).join('/')
 
-    return client(path, { method: 'PUT', body: data })
+    return client(path, { method: 'PUT', body: { data } })
   }
 
   /**
@@ -88,7 +78,6 @@ export const useStrapi3 = () => {
   }
 
   return {
-    count,
     find,
     findOne,
     create,
