+++ src/runtime/strapi.plugin.ts
@@ -1,4 +1,4 @@
-import { useStrapiAuth } from '../composables/useStrapiAuth'
+import { useStrapiAuth } from './composables/useStrapiAuth'
 import { defineNuxtPlugin } from '#app'
 
 export default defineNuxtPlugin(async () => {
